requirements.txt
requests
beautifulsoup4
lxml
flask
flask-cors
python-dotenv 